Extracted whole service logic between views.py and models.py to the separate layer services.py
+small typo fix
